adding basic test and few fixes
