+++ shoebox/app/com/keepit/common/mail/EmailAddresses.scala
@@ -8,12 +8,12 @@ trait EmailAddressHolder {
 sealed abstract class SystemEmailAddress(val address: String) extends EmailAddressHolder
 
 object EmailAddresses {
-  case object TEAM extends SystemEmailAddress("team@keepit.com")
-  case object SUPPORT extends SystemEmailAddress("support@keepit.com")
-  case object ENG extends SystemEmailAddress("eng@keepit.com")
-  case object EISHAY extends SystemEmailAddress("eishay@keepit.com")
-  case object YASUHIRO extends SystemEmailAddress("yasuhiro@keepit.com")
-  case object ANDREW extends SystemEmailAddress("andrew@keepit.com")
+  case object TEAM extends SystemEmailAddress("team@42go.com")
+  case object SUPPORT extends SystemEmailAddress("support@42go.com")
+  case object ENG extends SystemEmailAddress("eng@42go.com")
+  case object EISHAY extends SystemEmailAddress("eishay@42go.com")
+  case object YASUHIRO extends SystemEmailAddress("yasuhiro@42go.com")
+  case object ANDREW extends SystemEmailAddress("andrew@42go.com")
 
   def apply(email: String): SystemEmailAddress = email match {
     case TEAM.address => TEAM
